@@ -10,33 +10,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/cdk": "^19.2.18",
-    "@angular/common": "^19.2.0",
-    "@angular/compiler": "^19.2.0",
-    "@angular/core": "^19.2.0",
-    "@angular/forms": "^19.2.0",
-    "@angular/material": "^19.2.18",
-    "@angular/platform-browser": "^19.2.0",
-    "@angular/platform-browser-dynamic": "^19.2.0",
-    "@angular/router": "^19.2.0",
+    "@angular/animations": "^20.3.4",
+    "@angular/cdk": "^20.2.8",
+    "@angular/common": "^20.3.4",
+    "@angular/compiler": "^20.3.4",
+    "@angular/core": "^20.3.4",
+    "@angular/forms": "^20.3.4",
+    "@angular/material": "^20.2.8",
+    "@angular/platform-browser": "^20.3.4",
+    "@angular/platform-browser-dynamic": "^20.3.4",
+    "@angular/router": "^20.3.4",
     "@ngx-formly/core": "^7.0.0",
     "@ngx-formly/primeng": "^7.0.0",
     "@ngx-translate/core": "^16.0.4",
     "@ngx-translate/http-loader": "^16.0.1",
-    "@primeng/themes": "^19.1.3",
+    "@primeng/themes": "^20.2.0",
     "@tailwindcss/postcss": "^4.1.8",
     "chart.js": "^4.4.2",
     "primeicons": "^7.0.0",
-    "primeng": "^19.1.3",
+    "primeng": "^20.2.0",
     "rxjs": "~7.8.0",
     "tailwindcss-primeui": "^0.6.1",
-    "tslib": "^2.3.0",
+    "tslib": "^2.8.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.14",
-    "@angular/cli": "^19.2.14",
-    "@angular/compiler-cli": "^19.2.0",
+    "@angular/build": "^20.3.5",
+    "@angular/cli": "^20.3.5",
+    "@angular/compiler-cli": "^20.3.4",
     "@ngx-formly/schematics": "^6.3.12",
     "@schematics/angular": "^19.2.15",
     "@types/jasmine": "~5.1.0",
@@ -47,8 +48,8 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "postcss": "^8.5.6",
+    "postcss": "^8.5.3",
     "tailwindcss": "^4.1.11",
-    "typescript": "~5.7.2"
+    "typescript": "~5.9.3"
   }
 }
