@@ -4,7 +4,6 @@ import { TagModule } from 'primeng/tag';
 import { RatingModule } from 'primeng/rating';
 import { CommonModule } from '@angular/common';
 import { ButtonModule } from 'primeng/button';
-import { MessageService } from 'primeng/api';
 import { ToastModule } from 'primeng/toast';
 import { ProductService } from '@login/services/product.service';
 import { PackageComponent } from '@shared/pages/package/package.component';
@@ -17,11 +16,12 @@ import { DatePickerModule } from "primeng/datepicker";
 import { FormsModule } from '@angular/forms';
 import { InputTextModule } from 'primeng/inputtext';
 import { DoctorService } from '@operations/services/doctor.service';
+import { MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-packages',
   imports: [TableModule, TagModule, ToastModule, RatingModule, ButtonModule,FormsModule, InputTextModule, CommonModule, PackageComponent, IconFieldModule, InputIconModule, DatePickerModule],
-   providers: [ProductService, MessageService],
+   providers: [ProductService],
   templateUrl: './packages.component.html',
   styleUrl: './packages.component.scss'
 })
@@ -29,8 +29,9 @@ export class PackagesComponent {
 
   supplierService = inject(SupplierService)
   doctorService = inject(DoctorService)
-  messageService = inject(MessageService)
+  message = inject(MessageService)
   authService = inject(AuthService)
+
   products = signal([])
   role = this.authService.userRole
  @Input() filter :any= {}
@@ -50,7 +51,7 @@ DataChange(event){
      switch (event.name) {
       case 'buy':
         console.log(event.product);
-        this.addToCard(event.product)
+        this.addToCard(event.product , event.index)
         break;
 
       case 'delete order':
@@ -68,13 +69,23 @@ DataChange(event){
         this.loadData(event.filter)
         break;
     
+      case 'update':
+        console.log(event.product);
+        this.updateProduct(event.product ,event.order)
+        break;
+
+      case 'activate':
+        console.log(event.order);
+        this.activate(event.order)
+        break;
+    
       default:
         break;
     }
 }
 loadData(filter={}){
   this.filter.page = filter['page']
-    this.supplierService.getOrders(this.filter).subscribe((res:any) => {
+    this.supplierService.getPackages(this.filter).subscribe((res:any) => {
       console.log(res['data']);
         this.products.set(res['data']);
         this.totalRecords =res.meta.total
@@ -83,7 +94,7 @@ loadData(filter={}){
 markAsFavorite(product , index){
   this.doctorService.markAsFavorite(product.id).subscribe((res) => {
      product.favorite = true
-    this.messageService.add({
+    this.message.add({
           severity: 'success',
           summary: 'Successful',
           detail: 'Product added to favorites',
@@ -91,16 +102,16 @@ markAsFavorite(product , index){
     });
   });
 }  
-addToCard(product){
+addToCard(product,index){
   let data = {
-    product_id: product.id,
+  "notes": "يرجى التوصيل بعد الساعة 5 مساءً",
+  "payment_method": "مدفوعات" // "الكترونى" , "كاش" , مدفوعات
 }
-product.is_added = true
-   this.doctorService.addToCard(data).subscribe((res) => {
-    this.messageService.add({
+   this.doctorService.addPackageToCard(product.id ,data).subscribe((res) => {
+    this.message.add({
           severity: 'success',
           summary: 'Successful',
-          detail: 'Product added to Card',
+          detail: 'Package added to Current Order',
           life: 3000
     });
   });
@@ -128,7 +139,7 @@ onSearch() {
   this.searchSubject.pipe(
     debounceTime(300),
       // distinctUntilChanged(), 
-      switchMap((val) =>  this.supplierService.getOrders(this.changeFormat(val)))
+      switchMap((val) =>  this.supplierService.getPackages(this.changeFormat(val)))
     )
     .subscribe(res=>{
         this.products.set( res['data']);
@@ -136,15 +147,75 @@ onSearch() {
     })
 }
 deleteOrder(order,index){
-  this.supplierService.deleteOrder(order.id).subscribe(res=>{
+  this.supplierService.deletePackage(order).subscribe(res=>{
     this.products().splice(index,1)
+  this.message.add({
+        severity: 'success',
+        summary: 'Successful',
+        detail: 'Item Deleted Successfully',
+        life: 3000
+      });
   })
 }   
-updateProduct(product , oldProduct){
-  this.supplierService.updateOrderProduct(product).subscribe(res=>{
-      oldProduct.quantity = oldProduct.quantity - product.quantity
-    })
+updateProduct(product ,order){
+    // console.log(product);
+    // let x = {
+    // "id": 2,
+    // "doctor_name": "Doctor",
+    // "notes": "يرجى التوصيل بعد الساعة 5 مساءً",
+    // "status": "delivered",
+    // "status_name": "تم التوصيل",
+    // "total_order_price": 2546,
+    // "created_at": "2025-09-28",
+    // "products": [
+    //     {
+    //         "id": 3,
+    //         "product_id": 22,
+    //         "category_id": 6,
+    //         "category_name": "تقويم الأسنان",
+    //         "name": "hesham22",
+    //         "desc": "heshamskasdasd",
+    //         "img": "https://dental-link.azda-cs.com/storage/products/YIwubSDPtAyLaO1dZBwennFlZ0zFTAD9SgCJYET9.png",
+    //         "price": 1000,
+    //         "quantity": 2,
+    //         "quantityAvailable": "2",
+    //         "total_price": 2000
+    //     },
+    //     {
+    //         "id": 4,
+    //         "product_id": 21,
+    //         "category_id": 6,
+    //         "category_name": "تقويم الأسنان",
+    //         "name": "sdfd",
+    //         "desc": "asd",
+    //         "img": "https://dental-link.azda-cs.com/storage/products/4SvFTjMj8Sbe3PaWIBvGqkz4Hqs6Njf4LgTjfcKb.png",
+    //         "price": 546,
+    //         "quantity": 1,
+    //         "quantityAvailable": "56",
+    //         "total_price": 546
+    //     }
+    // ]
+    // }
+  // Object.assign(order,x)
+  // this.supplierService.updatePackage(product).subscribe(res=>{
+  //   })
 }   
+activate(order) {
+  this.supplierService.activate(order.id).subscribe({
+    next: (res) => {
+      this.message.add({
+        severity: 'success',
+        summary: 'Successful',
+        detail: 'Activate changed',
+        life: 3000
+      });
+    },
+    error: () => {
+      // لو حصل error رجّع الحالة القديمة
+      order.checked = !order.checked;
+    }
+  });
+}
 changeFormat(val){
   let obj = {
     id : val?.id,
